+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - [enhancement:] Callbacks for `add`, `cancel` & `cancelAll`
 - [enhancement:] `image:` accepts remote URLs and local URIs (Android)
 - [enhancement:] Schedule multiple notifications at once (Android)
-- [enhancement:] Cancel multiple notifications at once (Android)
+- [enhancement:] Cancel multiple notifications at once
 - [enhancement:] Clear multiple notifications at once (Android)
 - [enhancement:] `clear` & `clearAll` methods (Android)
 - [enhancement:] `onclear` event (Android)